+++ R/get_data_object.R
@@ -306,6 +306,13 @@ get_data_object_spautor <- function(formula, data, spcov_initial,
       if (inherits(spcov_initial, "sar")) {
         warning("M ignored for sar models", call. = FALSE)
       }
+      if (inherits(M, "matrix")) {
+        if (dim(M)[1] == dim(M)[2]) {
+          M <- diag(M) # take diagonal of matrix
+        } else {
+          M <- as.vector(M) # assume diagonal already given as one-column vector
+        }
+      }
     }
   }